gui: Prevent plugins from interfering with each other.
The GTK 2 and GTK 3 plugins cannot cannot coexist in the same process
so the corresponding plugins must be unloaded as soon as we have their
priority. But even unloading + reloading GTK 2 or GTK 3 can lead to
random crashes.
So hold the GUI plugin tryouts in a subprocess.
Since this is a platform-specific operation add the necessary
infrastructure in platform_{unix,windows}.c.
Reuse MojoSetup_loadGuiPlugin() to simplify initGuiPluginsByPriority().

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>