upgrade: Delete obsolete files and directories.
Lingering obsolete files (e.g. plugins) could break the upgraded
product including the postinstall script.
Furthermore obsolete files disappear from the manifest file which means
they are not deleted on uninstall either, resulting in unclean
uninstalls.
So load the old manifest and delete the obsolete files and directories
found there.
But be careful when loading the old Lua manifest so that a corrupted
file does not crash the installer and does not clobber our state
information.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>