stb_image: Fix some misleading 'for' indentation warnings.
gcc -Wmisleading-indentation does not like it when a 'break;' follows a
for's simple statement because that makes it look like the break is part
of the for.
In fact it does not quite look that way because the for is inside a
macro! Still, add braces around the for statement to avoid any
ambiguity.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>