gtk2&3: Change the minimum and default window size.
It seems that the only point of setting a minimum size for the license
area is to force a certain size on the overall window. So it's much
better to do that directly through gtk_window_set_default_size().
Also gdk_screen_{width,height}() are deprecated in GTK 3 and don't give
good results anyway. So simply use a hardcoded default window size that
gives good results based on the content of the window. Note that since
this is just a default, the window size will be adjusted anyway if it
is bigger than the desktop size, or to accomodate a large banner image.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>