Add support for the recommended_full_destinations attribute.
This is the same as the recommended_destinations attribute except that
the id attribute is not appended. This can be used to automatically
suggest installing in a non standard destination (i.e. one that does
not end with id) for an upgrade if the installer has a way to detecting
such install paths.
Sorting the candidate destinations (alphabetically, most recently used
path first, etc.) is the responsibility of the code setting the
attribute.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>