isfile() and isdir() should follow symbolic links.
This is what the equivalent operators / functions do in POSIX shells,
perl and Python. So doing differently is very confusing.
It's easy to do an issymlink() check where needed but much harder to
manually dereference the symbolic link (recursively) to see whether it
points to a file or directory after all.
Also the explicit issymlink() checks make the code much clearer.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>