upgrade: Fix rollbacks when a file or symlink replaces a directory.
First we must make sure not to move non-empty directories to the
rollback directory. Otherwise we would be unable to rmdir them when
clearing the rollbacks.
Furthermore we must make sure to mark all their content as seen so we
don't try to delete it as 'obsolete' later on. This is critical when a
directory is replaced by a symbolic link pointing to a directory that
may well contain the same set of files.
Note that this requires iterating on the content of directories and thus
requires access to opendir().

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>