gtk2&3: Treat closing the window as clicking Finish when possible.
Otherwise closing the window instead of clicking on the 'Finish' button
results in a dialog asking 'Are you sure you want to cancel the
installation?'. Clicking 'No' returns to the 'Finish' dialog but
clicking 'Yes' does not abort the installation since it is complete can
cannot be rolled back anymore.

From: Esme Povirk <esme@codeweavers.com>
Signed-off-by: Francois Gouget <fgouget@codeweavers.com>