Do the cleanups before presenting the final dialog.
Otherwise we can get a long freeze after everything should be finished
from the user's point of view.
This leaves the progress bar stuck at 100% but shows the file being
cleaned up so users see that something is being done. This allows
disabling the 'Cancel' button so the cleanup phase does not get
interrupted.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>